Automatically expand folds in Markdown
zR is the mapping to expand all folds, and since vim-markdown doesn't
have a specific setting for this we can run that normal command when
opening markdown files (this strategy was in a vim-markdown issue).

The other option just ensures Liquid front-matter gets highlighted and
is unrelated, but I couldn't split that hunk.