Change leader key from comma to space
Comma is a standard Vim mapping to repeat the latest f, t, F, or T
motion in the opposite direction. It's not really something I used
but overriding that seems like a questionable idea.

Plus space is a big fat target to hit. It allows you to easier hit
certain keys which is a big win on its own.