Prevent cursor from jumping up after yanking
When visually selecting lines, if you yank them then your cursor will
jump to where you started your visual selection but chances are you want
it to be at the bottom.

This bind changes the behavior of 'y' in visual selection mode to keep
it at the bottom of your selection after yanking.