Update Vim buffers if they change outside of Vim
In order for this to work autoread must be set and the file must be
saved before you perform the action that would modify the buffer
externally.

For example, if you're running a linting tool that will auto-format code
then you'll want to save your changes in the buffer before running that
auto-format tool, otherwise the buffer won't get loaded. That's smart
because otherwise you would have lost your unsaved changes.