Replace Markdown plugin to use tpope's
The plasticboy plugin was constantly bugging me with concealing headers
even when I turned that behavior off, and it also had other issues,
such as:

The indentation handling of bullet points wasn't optimal for me. It kept
indenting paragraphs after the last bullet.

Another issue is you can break syntax highlighting for the whole file
very easily with certain links. I opened an issue for this back in March
and a proposed fix was made shortly after but it hasn't been merged.

That issue is at: https://github.com/plasticboy/vim-markdown/issues/419

In the end, I am switching. Tim's plugin is maintained and I've been
using it for a bit now without issues.