Remove lazyredraw
Back when I first tried Vim on WSL, it was really slow to the point of
being unusable. It would take 5+ seconds to move your cursor even in
a tiny 100 line file with syntax highlighting.

I don't know what happened with WSL exactly, but now this problem is
gone and having this option removed is kind of nice because Vim will
update the screen more often. It makes Vim feel a bit more responsive at
the UI level after performing certain actions that cause a UI change.

I'm going to keep an eye on this but so far so good, even when dealing
with massive 100,000+ line Markdown files.