Add 24-bit color support and switch to One Dark
This commit has 2 changes intermingled in it.

This new strategy to use 24-bit colors in the vim / tmux configs means
that shell script is no longer necessary to source in your ~/.profile.

This is great news because it means switching color schemes is much easier
now. You can simply install your new theme and use it with no extra fuss.

The other change is switching from Gruvbox to One Dark. A future commit
will update the README file with reasons why along with new screenshots.