Add quick-scope Vim plugin
I'm starting to use f more often to jump to specific points in a line
and this plugin adds a visual cue on how to quickest jump to the
character. Very useful!

This visual cue is only shown when you press one of the horizontal
movement keys while being on a line with the settings used here.