Enable Vim's autocomplete menu
This does not require having any language servers installed. It will
bring up the complete menu automatically thanks to the plugin and then
the custom binds allow you to interact with the menu.

Of course you can change those arrow keys to use hjkl if you prefer.

The only thing the plugin does is allow for the complete menu to pop up
without having to hit CTRL+p in insert mode, so if you don't like the
automated behavior you can easily remove that.