Add custom settings for vim-test
The executables are set up to expect that you want to run your tests
inside of a Docker container where the Docker Compose service name is
"web".

This works in all of my projects because I keep that service name the
same between projects. Your mileage may vary.

The keybinds let you control which tests to run.