Change the tab complete custom bind to enter
This wasn't working in a way that I could get used to using. Plus it
started to mess with using the tab key when the menu wasn't shown.

Now it's been changed to hitting enter to complete. This works 95% of
the time. The only time it really fails is when you want to not complete
something and immediately hit enter to move to the next line, in which
case you need to remember to cancel the completion (left arrow) before
hitting enter.

I'm finding this doesn't happen often, but yeah, if anyone knows how to
100% flawlessly handle tab feel free to open a PR!