Support both WSL 1 and WSL 2
Unfortunately using :0 as the DISPLAY no longer works on WSL 2 at the
time of making this commit. Perhaps that will change in the future.

Having used WSL 2 for only 1 day now, I can safely say that as of today
I will be still using WSL 1 for when I want to grep and manipulate files
in my mounted drives. The performance is a billion times better for that.