Update tmux config for 3.0+
It's quite easy to get 3.0+ on most distros of Linux nowadays. Sorry in
advance if you're using an older version.

If a large demand for supporting 2.x comes in, there are options to
situationally set specific variables based on the current tmux version.

An example of that can be found here: https://stackoverflow.com/a/40902312