Add VcXsrv config to Windows start up
I've had this on my machine forever, but forgot to port it back to my
dotfiles repo. This is super handy for never having to worry about
configuring this again.

It comes up when Windows boots and clipboard sharing works.

Of course you'll want to replace "Nick" with your user name when you
create this file on your system.