Add Microsoft Terminal settings.json file
Truthfully I'm not sure if that 8weky ID is unique to my machine or not. It
comes up in Google searches for other applications.

In any case, if that's not your path just go to the Packages directory
and find your WindowsTerminal directory. Hopefully Microsoft makes it
easier to share this config file in the future.