Switch wsltty font from Consolas NF to Consolas
Also just a heads up, I'll likely be removing this entire wsltty
directory from my dotfiles in the future.

Leaving it in for now because I still very much trust it, where as the
Microsoft Terminal only has a few weeks of real world usage under its
belt. It still has weird color flicking issues once in a while but
nothing that is a deal breaker.

Now with WSL 2 being out, the Microsoft Terminal's ability to quickly
launch a WSL 1 or WSL 2 terminal is a huge quality of life improvement.