Add file preview to custom fzf Rg command
I had no idea this existed until today when I was disabling a few things
in my vimrc and noticed that without this the fzf Rg command would show
a file preview on the side. So cool.

Now I've updated things to use custom rg flags and still have the file
preview ability.