Add protection against WSL 2 using all your memory
At the time of writing this commit there's a pending issue for WSL 2
where it happily eats all of your memory over time.

Details are at: https://github.com/microsoft/WSL/issues/4166

I ran into this issue naturally starting and stopping Docker containers
for half a day while doing my normal work. Before I knew it, the VM was
using 11GB of memory (the default is to use 80% of your RAM).

I'm not going to bikeshed on this by making a cronjob run the alias
command on a schedule because I imagine this will be fixed in a hotfix
at some point in the future.

Just be mindful of this if you notice that the Vmmem process is using a
ton of memory.