Change Vim theme to use vim-one
This is still basically the same as the old theme, except this version
of it supports both a dark and light variant which I really like.

There's a few little changes for some the colors with this version of
One vs the previous theme. The only thing I'm not crazy about is when
you have multiple Vim buffers open the active buffer's status bar is
dimmer than the inactive one.

That is something I may change in the future for this theme but for now
it's not too much of a distraction. I'm more concerned about it for when
I create video tutorials. If I ever split a file, folks may think the
active one is inactive and vice versa.