Add explicit configuration for fzf colors
By default it looked good on dark terminal and Vim themes, but if
you switch to a light theme it looked out of place.

You can toggle this to be dark, light or a number of other different
values. You can even set specific colors if you don't like the default
colors for dark and light.

Details are at: https://github.com/junegunn/fzf/wiki/Color-schemes