Add toggle dark / light mode script
It was starting to get tedious to manually switch between dark and light
mode in a few different terminal tools.

If you want to use this script on your end, don't forget to change your
terminal config path to match where it exists on your machine along with
any config paths or search / replace strings to match your theme's name
and colors.