Add optional aliases file for custom aliases
Some aliases are useful for multiple folks but I wanted to be able to
add personal aliases that are hyper specific to me.

For example, an alias to access specific files or directories that won't
exist on your machine such as contract related work.

This pattern allows an ability to add personalized aliases that won't
make their way into version control.