Fix a few configs losing their symlink
Turns out the perl command being used to do inplace edits also
transformed the symlink into a real file.

I couldn't find a solution that seems compatible on Linux and MacOS
which allowed the perl command to not clobber the symlink.

So it's back to using sed with a stipulation that if you're on MacOS
you'll want to install the GNU version of sed and also create an alias
so that running sed uses gsed.