Fix MS Terminal config file from getting corrupt
Getting rid of the stderr=PIPE alone didn't fix it, but adding the sleep
along with removing the stderror option does.

I'm still not 100% sure if it's fixed but I let this script run in a
while loop with a 100ms delay for 45 minutes straight and it never got
into a corrupt state.

Without this patch it would have gotten corrupt within 15 seconds or so
and I was able to reproduce this semi-regularly, probably 80% of the
time it would happen within 15 seconds and 100% within 1 minute.