Prevent from Vim autoscrolling down 3 lines
I had this setting for so long but really it's been causing me more
grief than anything.

This reverts this setting back to its default value. Now Vim won't
scroll down 3 lines from the top of the buffer if you happen to be
working on a line near the top of your visible screen.