Remap new keys for moving lines up and down
Using J (shift-j) to move a line up to the previous line was something I
did all the time and the old mappings removed that.

These new mappings shouldn't conflict with anything and now we have new
mappings to use the arrow keys too.

This also required configuring tmux to use xterm compatible control
arrow keys to get this to work in Vim.