Really fix history for recording aliases
The problem with the old version was if I opened another terminal, the
history wasn't being restored.

This now works almost the same as it did with Bash in the sense that if you
start recording in one terminal, the history is cleared in that terminal
and all future terminals you open.

Then if you stop recording it brings your history back in the current
terminal along with any new terminals you open. However if you stopped
recording in a 2nd terminal and your first terminal is still open,
you'll need to close that first terminal since it won't have any history.