Add iproute2 to installer's apt packages
This isn't installed inside of Docker which makes locally testing this
script a bit annoying. It should be installed by default in most major
Debian based distros of Linux so it's not adding unnecessary bloat.

It needs to exist anyways because it's being used in my .zshrc to
configure the DISPLAY inside of WSL 2.