Use cp instead of mv for history manipulation
Using mv is risky. After using these scripts about 100 times I ended up
in a scenario where I accidentally moved an empty backup to be my main
history which resulted in losing my history.

I had a daily backup from another drive so it wasn't the end of the
world but it makes me think that I'd rather use cp here instead of mv to
ensure this doesn't happen again.

I'm even going as far as keeping the lingering .bak file around after
stopping the recording. This won't be used as a source of truth because it
only  gets updated in between recording sessions but this feels worth it
to leave around as an insurance policy in case something crazy happens.