Fix current shell history in start/stop-record
Since I moved everything into their own dedicated shell scripts, running
commands in the script to modify the current shell didn't work anymore.

Now I've created new aliases so I can clear the history in my current
shell so that my current shell's history is empty right after running
the start-rec alias.

This has an added benefit of making the dedicated scripts more portable
because now they work the same in bash and zsh since the current shell
related bits are moved to an alias. In this case both bash and zsh
support setting 0 as the HISTSIZE so the alias works in both shells.

However the rm command had issues with zsh when there were no tmp files
to delete due to how it handles the star. That problem is gone now that
the script has been set back to use bash.