Use the default dotfiles path for dt
This works for more folks out of the box. In my case I still have it in
the ~/src/github/dotfiles location but I overwrote this alias in the
.aliases.local file so we have the best of both worlds.

I also removed a few aliases I rarely use. Since everything is now in
the dotfiles repo, having this single alias is enough because you can
quickly jump to any of those other files inside of Neovim.