Fix .env.example being set as conf
Turns out ordering is not guaranteed, I was getting non-deterministic
results with .env.example being set as dotenv or conf randomly.

Priority is supported which is documented. The higher priority wins and
if it's not defined it defaults to 0. We want .env.example last because
the first pattern matches everything.