Fix local zsh files when running in Docker
If you had a previously defined .zprofile.local file with a
DOTFILES_PATH set then it would use that in Docker and that path likely
wouldn't exist.

Most folks won't run into this issue, it's mainly for isolated tests but
now it's fixed in a way that should always work.