Document how to use a right aligned explorer
Some folks prefer this and it wasn't easy to find how to do this in the
docs. The reason I don't use a right aligned explorer is because then
the explorer is very very far away from your code when you have a large
screen resolution and a full size terminal window.