Move zsh history into dotfiles (gitgnored)
The .cache/ directory wasn't a great spot because it could be
accidentally cleared and losing your history would be bad.

I almost lost 2 years of history because of this. Thankfully I had a
backup from a few hours before I deleted my .cache/ directory.

Now this file will be straight up included in your dotfiles and
gitignored. Since it's gitignored, the Snacks picker and Grug will both
avoid referencing this file unless you toggle it on.