Ensure certain files get deleted in a container
This is to help make sure when testing this script we have a fresh
environment without any volume mounted in local files or history.

We would never want to delete these files when you run the script on
your normal system. It would be really random if you happen to have
IN_CONTAINER=1 set in your shell by default on your real system.