Add Ghostty support and drop iTerm2 support
I got a chance to try Ghostty on macOS recently. It's really solid.

On a work laptop with macOS I will be using Ghostty, I would on native
Linux too.

With that said, I'll entertain supporting other terminals in these
dotfiles if folks want to open a PR but in the case of iTerm2 there were
a number of manual steps to take to configure it or switch themes so I
don't see a compelling reason to support it anymore.