Automate installing Inconsolata Nerd Font
I almost got it fully automated on Windows but unfortunately even if you
copy files from WSL 2 into Windows successfully, Windows would not
accept the font as being installed correctly.

The Microsoft Terminal would hard crash when trying to load with the
font referenced. It found a reference to the font but it was like the
file itself wasn't valid.