Add cross OS clipboard copy / paste scripts
This is mainly to make it easier to copy / paste within scripts
contained in this dotfiles repo.

But, as someone who uses Linux on my personal machine but a Mac for work
sometimes, it's nice to have 1 unified command to run too. If you ever
need to reach for more specific features available for 1 specific tool
you can do that as needed.