Replace pass with keepassxc
I had a good run with pass but syncing passwords across multiple
device is much easier with keepassxc since it's 1 DB file.

It took me ~3 hours to manually convert almost 400 passwords over, it is
a one time pain. It could have been faster with a script to import
things but I took this as an opportunity to audit things too.