Add bootstrap check for empty DOTFILES_PATH
This will be the case for the first installation where DOTFILES_PATH
will be empty since it gets set later in the script.

With that said, it's likely safe to assume the script is being called
from the BOOTSTRAP_PATH here.

If this ends up being a problem it can get its own dedicated elif and we
can try to use the relative ./install-config path.