Support cloning and updating from a custom branch
I'm preparing to update a ton of things, such as adding niri support and
other graphical tools. This will be opt-in.

I want to be able to push a PR up and provision against that instead of
the master branch.

Needless to say, being able to test PRs is important!